Oops, these days, magic numbers are supposed to be type "long", and the m4i64 machines really do care about this!
svn:revision:14709
cvs:account:mike
cvs:branch:trunk

+++ h/shadefuncs.h
@@ -15,7 +15,7 @@
  *  The interface to the various material property & texture routines.
  */
 struct mfuncs {
-	int		mf_magic;	/* To validate structure */
+	long		mf_magic;	/* To validate structure */
 	char		*mf_name;	/* Keyword for material */
 	struct mfuncs	*mf_forw;	/* Forward link */
 	int		mf_inputs;	/* shadework inputs needed */
